Fix CVE-2017-7467, a remote exploitalbe out of bound access
This fixes an out of bounds data access that can lead to a
remotely exploitable code execution.

This issue was found by Solar Designer of Openwall during a
security audit of the Virtuozzo 7 product, which contains
derived downstream code in its prl-vzvncserver component.
The corresponding Virtuozzo 7 fix is:

https://src.openvz.org/projects/OVZ/repos/prl-vzvncserver/commits/6d95404e75b98f36b1cc85ee23df99dcf06ca13f

Openwall would like to thank the Virtuozzo company for
funding the effort.